requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://mirrors.aliyun.com/pypi/simple
allure-pytest==2.8.18
allure-python-commons==2.8.18
attrs==21.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
certifi==2020.12.5
chardet==3.0.4
idna==2.8
importlib-metadata==4.0.1; python_version < '3.8'
iniconfig==1.1.1
more-itertools==8.7.0; python_version >= '3.5'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.10.1
pymysql==0.9.3
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest==6.0.1
pytest-ordering==0.6
redis==3.5.3
requests==2.22.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
xlrd==1.2.0
zipp==3.4.1; python_version >= '3.6'
dict==2020.12.3

